# scripts/openapi.yaml
openapi: '3.1.0'
info:
  title: Visualizations of Security Log Data
  description: Some maps showing the origins of breakin attempts to my servers
  contact:
    name: Daniel Boline
    email: ddboline@gmail.com
  license:
    name: ''
  version: '0.13.2'
paths:
  /security_log/cleanup:
    post:
      operationId: host_country_cleanup
      responses:
        '201':
          description: Host Country Cleanup
          content:
            text/html:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HostCountry'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    type: string
        '401':
          description: Not Authorized
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    type: string
  /security_log/host_country:
    get:
      operationId: host_country_get
      parameters:
      - name: offset
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          minimum: 0
      - name: limit
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          minimum: 0
      responses:
        '200':
          description: Host Countries
          content:
            text/html:
              schema:
                type: object
                required:
                - pagination
                - data
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/HostCountry'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    type: string
        '401':
          description: Not Authorized
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    type: string
    post:
      operationId: host_country_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HostCountryUpdate'
        required: true
      responses:
        '201':
          description: Host Country Post
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    type: string
        '401':
          description: Not Authorized
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    type: string
  /security_log/intrusion_attempts:
    get:
      operationId: intrusion_attempts
      parameters:
      - name: service
        in: query
        required: false
        schema:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/Service'
      - name: location
        in: query
        required: false
        schema:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/Host'
      - name: ndays
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          format: int32
      responses:
        '200':
          description: Intrusion Attempts
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    type: string
        '401':
          description: Not Authorized
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    type: string
  /security_log/intrusion_attempts/all:
    get:
      operationId: intrusion_attempts_all
      parameters:
      - name: service
        in: query
        required: false
        schema:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/Service'
      - name: location
        in: query
        required: false
        schema:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/Host'
      - name: ndays
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          format: int32
      responses:
        '200':
          description: All Intrusion Attempts
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    type: string
        '401':
          description: Not Authorized
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    type: string
  /security_log/intrusion_log:
    get:
      operationId: intursion_log_get
      parameters:
      - name: service
        in: query
        required: false
        schema:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/Service'
      - name: server
        in: query
        required: false
        schema:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/Host'
      - name: offset
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          minimum: 0
      - name: limit
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          minimum: 0
      responses:
        '200':
          description: Intrusion Logs
          content:
            text/html:
              schema:
                type: object
                required:
                - pagination
                - data
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/IntrusionLog'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    type: string
        '401':
          description: Not Authorized
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    type: string
    post:
      operationId: intrusion_log_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IntrusionLogUpdate'
        required: true
      responses:
        '201':
          description: Intrusion Log Post
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    type: string
        '401':
          description: Not Authorized
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    type: string
  /security_log/log_messages:
    get:
      operationId: get_log_messages
      parameters:
      - name: log_level
        in: query
        required: false
        schema:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/LogLevel'
      - name: log_unit
        in: query
        required: false
        schema:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/str'
      - name: min_date
        in: query
        required: false
        schema:
          type:
          - string
          - 'null'
          format: date-time
      - name: max_date
        in: query
        required: false
        schema:
          type:
          - string
          - 'null'
          format: date-time
      - name: limit
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          minimum: 0
      - name: offset
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          minimum: 0
      responses:
        '200':
          description: Log Messages
          content:
            text/html:
              schema:
                type: object
                required:
                - pagination
                - data
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/SystemdLogMessages'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    type: string
        '401':
          description: Not Authorized
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    type: string
...
components:
  schemas:
    Host:
      type: string
      enum:
      - Home
      - Cloud
    HostCountry:
      type: object
      required:
      - host
      - code
      - created_at
      properties:
        code:
          $ref: '#/components/schemas/str'
        created_at:
          type: string
          format: date-time
        host:
          $ref: '#/components/schemas/str'
        ipaddr:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/str'
    HostCountryUpdate:
      type: object
      required:
      - updates
      properties:
        updates:
          type: array
          items:
            $ref: '#/components/schemas/HostCountry'
    IntrusionLog:
      type: object
      required:
      - id
      - service
      - server
      - datetime
      - host
      properties:
        datetime:
          type: string
          format: date-time
        host:
          $ref: '#/components/schemas/str'
        id:
          type: string
          format: uuid
        server:
          $ref: '#/components/schemas/str'
        service:
          $ref: '#/components/schemas/str'
        username:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/str'
    IntrusionLogUpdate:
      type: object
      required:
      - updates
      properties:
        updates:
          type: array
          items:
            $ref: '#/components/schemas/IntrusionLog'
    LogLevel:
      type: string
      enum:
      - Debug
      - Info
      - Warning
      - Error
...
    Pagination:
      type: object
      required:
      - total
      - offset
      - limit
      properties:
        limit:
          type: integer
          minimum: 0
        offset:
          type: integer
          minimum: 0
        total:
          type: integer
          minimum: 0
    Service:
      type: string
      enum:
      - Apache
      - Nginx
      - Ssh
...
    str:
      type: string